requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.1

pip~=22.0.4
wheel~=0.36.2
setuptools~=57.0.0
zipp~=3.8.0